feat: changed BIOBANKER role name to REPRESENTATIVE to be more general purpose
