+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/dto/post/PostDTO.java
@@ -1,10 +1,13 @@
-package eu.bbmri.eric.csit.service.negotiator.api.dto.post;
+package eu.bbmri.eric.csit.service.negotiator.dto.post;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import eu.bbmri.eric.csit.service.negotiator.dto.person.PersonDTO;
 import eu.bbmri.eric.csit.service.negotiator.database.model.Attachment;
+import java.time.LocalDateTime;
 import java.util.Set;
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,18 +20,29 @@ import lombok.Setter;
 @AllArgsConstructor
 @Builder
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
-public class PostCreateDTO {
+public class PostDTO {
 
   Set<Attachment> attachments;
+  @NotNull
+  private String id;
+  @Valid
+  @NotEmpty
+  private String status;
   @Valid
   @NotEmpty
   private String text;
-  @Valid
-  @NotEmpty
-  private String resourceId;
 
   @Valid
-  private String status;
+  @NotEmpty
+  private LocalDateTime creationDate;
+
+  @Valid
+  @NotEmpty
+  private LocalDateTime modifiedDate;
+
+  @Valid
+  @NotEmpty
+  private PersonDTO poster;
 
 
 }
