+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/service/PostService.java
@@ -1,7 +1,7 @@
 package eu.bbmri.eric.csit.service.negotiator.service;
 
-import eu.bbmri.eric.csit.service.negotiator.api.dto.post.PostCreateDTO;
-import eu.bbmri.eric.csit.service.negotiator.api.dto.post.PostDTO;
+import eu.bbmri.eric.csit.service.negotiator.dto.post.PostCreateDTO;
+import eu.bbmri.eric.csit.service.negotiator.dto.post.PostDTO;
 import java.util.List;
 
 public interface PostService {
