+++ src/main/resources/data-postgres.sql
@@ -2,7 +2,8 @@ insert into role (id, name) values
   (1, 'CREATOR'),
   (2, 'ADMINISTRATOR'),
   (3, 'MANAGER'),
-  (4, 'RESEARCHER');
+  (4, 'RESEARCHER'),
+  (5, 'REPRESENTATIVE');
 
 insert into data_source (url, api_username, api_password, api_type, api_url, description, name,
                          resource_biobank, resource_collection, resource_network, source_prefix, sync_active)
